backend/requirements.txt
alembic==1.4.1
astroid==2.2.5
Babel==2.8.0
certifi==2019.11.28
cffi==1.14.0
Click==7.0
cryptography==43.0.3 # Last version that works with pyhton 3.7
ecdsa==0.13.2
Flask==1.0.2
Flask-Cors==3.0.8
flask-heroku==0.1.9
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.3
future==0.17.1
gunicorn==20.0.4
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.0
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
migrate==0.3.8
psycopg2-binary==2.8.4
pycparser==2.20
pycryptodome==3.3.1
pylint==2.3.1
pyOpenSSL==19.1.0
python-dateutil==2.6.0
python-editor==1.0.4
python-jose-cryptodome==1.3.2
pytz==2019.3
six==1.12.0
SQLAlchemy==1.3.3
typed-ast==1.3.5
Werkzeug==0.15.2
wrapt==1.11.1
python-dotenv==1.0.1
WTForms==2.2.1
